+++ Source/Alternet.UI/Controls/Grid/DoubleUtil.cs
@@ -9,9 +9,9 @@ namespace Alternet.UI
 {
     internal static class DoubleUtil
     {
-        // Const values come from sdk\inc\crt\float.h
+        // Const values come from sdk\inc\crt\double.h
         internal const double DBL_EPSILON = 2.2204460492503131e-016; /* smallest such that 1.0+DBL_EPSILON != 1.0 */
-        internal const float FLT_MIN = 1.175494351e-38F; /* Number close to zero, where float.MinValue is -float.MaxValue */
+        internal const double FLT_MIN = 1.175494351e-38F; /* Number close to zero, where double.MinValue is -double.MaxValue */
 
         /// <summary>
         /// AreClose - Returns whether or not two doubles are "close".  That is, whether or
