+++ Source/Alternet.UI/Handlers/Layout/StackPanel/StackPanelHandler.OrientedLayout.cs
@@ -14,14 +14,14 @@ namespace Alternet.UI
 
             public class AlignedPosition
             {
-                public AlignedPosition(float origin, float size)
+                public AlignedPosition(double origin, double size)
                 {
                     Origin = origin;
                     Size = size;
                 }
 
-                public float Origin { get; }
-                public float Size { get; }
+                public double Origin { get; }
+                public double Size { get; }
             }
 
             public StackPanel Control { get; }
