+++ Source/Alternet.UI/Native/Pen.cs
@@ -18,7 +18,7 @@ namespace Alternet.UI.Native
         {
         }
         
-        public void Initialize(PenDashStyle style, Alternet.Drawing.Color color, float width)
+        public void Initialize(PenDashStyle style, Alternet.Drawing.Color color, double width)
         {
             CheckDisposed();
             NativeApi.Pen_Initialize_(NativePointer, style, color, width);
@@ -34,7 +34,7 @@ namespace Alternet.UI.Native
             public static extern IntPtr Pen_Create_();
             
             [DllImport(NativeModuleName, CallingConvention = CallingConvention.Cdecl)]
-            public static extern void Pen_Initialize_(IntPtr obj, PenDashStyle style, NativeApiTypes.Color color, float width);
+            public static extern void Pen_Initialize_(IntPtr obj, PenDashStyle style, NativeApiTypes.Color color, double width);
             
         }
     }
