+++ Source/Alternet.UI/Utilities/StringTokenizer.cs
@@ -108,21 +108,21 @@ namespace Alternet.UI
             return result;
         }
 
-        public bool TryReadSingle(out float result, char? separator = null)
+        public bool TryReadSingle(out double result, char? separator = null)
         {
             if (TryReadString(out var stringResult, separator) &&
-                float.TryParse(stringResult, NumberStyles.Float, _formatProvider, out result))
+                double.TryParse(stringResult, NumberStyles.Float, _formatProvider, out result))
             {
                 return true;
             }
             else
             {
-                result = default(float);
+                result = default(double);
                 return false;
             }
         }
 
-        public float ReadSingle(char? separator = null)
+        public double ReadSingle(char? separator = null)
         {
             if (!TryReadSingle(out var result, separator))
             {
